Implement response DTO for GET /timezones endpoint
