+++ src/main/java/teammates/ui/webapi/action/GetTimeZonesAction.java
@@ -7,7 +7,7 @@ import java.util.Map;
 import java.util.TreeMap;
 
 import teammates.common.exception.UnauthorizedAccessException;
-import teammates.ui.webapi.output.ApiOutput;
+import teammates.ui.webapi.output.TimeZonesData;
 
 /**
  * Action: get supported time zones.
@@ -37,31 +37,7 @@ public class GetTimeZonesAction extends Action {
                 tzOffsets.put(tz, offset);
             }
         }
-        TimezoneData output = new TimezoneData(tzVersion, tzOffsets);
+        TimeZonesData output = new TimeZonesData(tzVersion, tzOffsets);
         return new JsonResult(output);
     }
-
-    /**
-     * Output format for {@link GetTimeZonesAction}.
-     */
-    public static class TimezoneData extends ApiOutput {
-
-        private final String version;
-        private final Map<String, Integer> offsets;
-
-        public TimezoneData(String version, Map<String, Integer> offsets) {
-            this.version = version;
-            this.offsets = offsets;
-        }
-
-        public String getVersion() {
-            return version;
-        }
-
-        public Map<String, Integer> getOffsets() {
-            return offsets;
-        }
-
-    }
-
 }
