+++ src/test/java/teammates/test/cases/webapi/GetTimeZonesActionTest.java
@@ -1,14 +1,12 @@
 package teammates.test.cases.webapi;
 
-import java.util.Map;
-
 import org.apache.http.HttpStatus;
 import org.testng.annotations.Test;
 
 import teammates.common.util.Const;
 import teammates.ui.webapi.action.GetTimeZonesAction;
-import teammates.ui.webapi.action.GetTimeZonesAction.TimezoneData;
 import teammates.ui.webapi.action.JsonResult;
+import teammates.ui.webapi.output.TimeZonesData;
 
 /**
  * SUT: {@link GetTimeZonesAction}.
@@ -28,25 +26,27 @@ public class GetTimeZonesActionTest extends BaseActionTest<GetTimeZonesAction> {
     @Override
     @Test
     protected void testExecute() throws Exception {
-
         ______TS("Normal case");
 
         GetTimeZonesAction a = getAction();
         JsonResult r = getJsonResult(a);
 
-        assertEquals(HttpStatus.SC_OK, r.getStatusCode());
 
+        TimeZonesData output = (TimeZonesData) r.getOutput();
+
+        assertEquals(HttpStatus.SC_OK, r.getStatusCode());
         // This test does not check the timezone database used is the latest
         // Only check that the version number is returned, and some sample values for timezone offset
-
-        TimezoneData output = (TimezoneData) r.getOutput();
-        Map<String, Integer> offsets = output.getOffsets();
         assertNotNull(output.getVersion());
-        assertEquals(8 * 60 * 60, offsets.get("Asia/Singapore").intValue());
-        assertEquals(-5 * 60 * 60, offsets.get("America/New_York").intValue());
-        assertEquals(11 * 60 * 60, offsets.get("Australia/Sydney").intValue());
-        assertEquals(0, offsets.get("Europe/London").intValue());
 
+        /* TODO the asserts below are brittle as the expected values are not guaranteed to be correct
+         * e.g. New York observes DST, so the offset is not always UTC-05:00 the entire year.
+         * e.g. timezones can change, like Caracas modifying their timezone. This affects the offset as well.
+         */
+        assertEquals(8 * 60 * 60, output.getOffsets().get("Asia/Singapore").intValue());
+        assertEquals(-5 * 60 * 60, output.getOffsets().get("America/New_York").intValue());
+        assertEquals(11 * 60 * 60, output.getOffsets().get("Australia/Sydney").intValue());
+        assertEquals(0, output.getOffsets().get("Europe/London").intValue());
     }
 
     @Override
@@ -54,5 +54,4 @@ public class GetTimeZonesActionTest extends BaseActionTest<GetTimeZonesAction> {
     protected void testAccessControl() throws Exception {
         verifyOnlyAdminCanAccess();
     }
-
 }
