+++ src/web/app/pages-admin/admin-timezone-page/admin-timezone-page.component.ts
@@ -1,11 +1,7 @@
 import { Component, OnInit } from '@angular/core';
 import { HttpRequestService } from '../../../services/http-request.service';
 import { TimezoneService } from '../../../services/timezone.service';
-
-interface TimezoneData {
-  version: string;
-  offsets: { [key: string]: number };
-}
+import { TimeZones } from "../../../types/api-output";
 
 /**
  * Timezone listing page for admin use.
@@ -27,7 +23,7 @@ export class AdminTimezonePageComponent implements OnInit {
   ngOnInit(): void {
     this.momentTzVersion = this.timezoneService.getTzVersion();
     this.momentTimezones = this.timezoneService.getTzOffsets();
-    this.httpRequestService.get('/timezone').subscribe((res: TimezoneData) => {
+    this.httpRequestService.get('/timezone').subscribe((res: TimeZones) => {
       this.javaTzVersion = res.version;
       this.javaTimezones = res.offsets;
     });
